requirements.txt
pandas
numpy
google-cloud-storage
scikit-learn
pyyaml
lightgbm
imbalanced-learn
mlflow # to keep track of models and experiments